Adicionando e disparando workflow via pr

+++ src/profile/profile.service.ts
@@ -1,20 +1,29 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { Profile } from './profile.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CreateProfileDto } from './dto/create-profile.dto';
+import { PaginationQueryDto } from 'src/common/pagination/dto/pagination-query.dto';
+import { Paginated } from 'src/common/pagination/pagination.interface';
+import { PaginationProvider } from 'src/common/pagination/pagination.provider';
 
 @Injectable()
 export class ProfileService {
   constructor(
     @InjectRepository(Profile)
     private readonly profileRepository: Repository<Profile>,
+    private readonly paginationProvider: PaginationProvider,
   ) {}
 
-  public async getAllProfiles() {
-    return this.profileRepository.find({
-      relations: ['user'],
-    });
+  public async getAllProfiles(
+    paginationDto: PaginationQueryDto,
+  ): Promise<Paginated<Profile>> {
+    return await this.paginationProvider.paginateQuery(
+      paginationDto,
+      this.profileRepository,
+      undefined,
+      ['user'],
+    );
   }
 
   public async createProfile(profileDto: CreateProfileDto) {
@@ -22,15 +31,15 @@ export class ProfileService {
     return await this.profileRepository.save(newProfile);
   }
 
-  public async updateProfile(id: number, profileDto: CreateProfileDto) {
+  public async updateProfile(userId: number, profileDto: CreateProfileDto) {
     // perfil existe para esse usuário?
     const profile = await this.profileRepository.findOne({
-      where: { user: { id } },
+      where: { user: { id: userId } },
       relations: ['user'],
     });
 
     if (!profile) {
-      return 'Profile not found';
+      throw new NotFoundException('This profile does not exist');
     }
 
     Object.assign(profile, profileDto);
@@ -38,6 +47,12 @@ export class ProfileService {
   }
 
   public async deleteProfile(id: number) {
+    const profile = await this.profileRepository.findOneBy({ id });
+
+    if (!profile) {
+      throw new NotFoundException(`This profile with id ${id} does not exist`);
+    }
+
     return await this.profileRepository.delete(id);
   }
 }
